% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROBIN.R
\name{ROBIN}
\alias{ROBIN}
\title{ROBIN (ROBust INitialization)}
\usage{
ROBIN(D, data, k, mp = 10, critRobin = 1.05)
}
\arguments{
\item{D}{A distance matrix calculated on \code{data}.}

\item{data}{A data matrix with n observations and p variables.}

\item{k}{The number of cluster centers to find.}

\item{mp}{The number of the nearest neighbors to find dense regions by LOF, the default is 10.}

\item{critRobin}{The cutoff value for LOF to determine the observations in the dense regions,
the default is 1.05.}
}
\value{
\item{centers}{A numeric vector of \code{k} initial cluster centers corresponding to the k indices of observations.}
\item{lof}{A real vector of local outlier factor values.}
}
\description{
\code{ROBIN} searches for k initial cluster seeds for k-means-based clustering methods.
}
\details{
The centers are the observations located in the most dense region and far away from each other at the same time.
In order to find the observations in the highly dense region, ROBIN uses LOF
(Local Outlier Factor, Breunig et al (2000)), see more details \code{\link[dbscan]{lof}}.
}
\examples{
# generate data
d <- SimData(size_grp=c(40,40,40),p_inf=50,
p_noise=750,p_out_noise=75)
dat <- scale(d$x)

res <- ROBIN(D=dist(dat),data=dat,k=3)
km <- kmeans(x=dat,centers=dat[res$centers,])
table(d$y,km$cluster)

}
\references{
Hasan AM, et al. Robust partitional clustering by
outlier and density insensitive seeding. Pattern Recognition Letters, 30(11), 994-1002, 2009.
}
\seealso{
\code{\link[dbscan]{lof}}, \code{\link[wrsk]{wrk}}
}
\author{
Sarka Brodinova <sarka.brodinova@tuwien.ac.at>
}
