% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimData.R
\name{SimData}
\alias{SimData}
\title{SimData}
\usage{
SimData(size_grp, p_inf, p_noise = NULL, p_out_inf = NULL, pct_out = 0.1,
  scatter_out = TRUE, p_out_noise = NULL, noise_pct_out = 0.1,
  unif_out_range = NULL, mu_grp_range = NULL, s_out_range = NULL,
  rho_grp_range = NULL)
}
\arguments{
\item{size_grp}{A numeric vector containing the group sizes to be generated. The length of the vector
corresponds to the number of groups to be generated.}

\item{p_inf}{The number of informative variables describing the generated group structure.}

\item{p_noise}{The number of noise variables which do not contribute to the group separation. If
not specified, no noise variables are generated.}

\item{p_out_inf}{The number of informative variables in which the observations are contaminated,
i.e. replaced by outliers either scatter outliers or uniformly distributed outliers, see \code{scatter_out}.}

\item{pct_out}{The proportion of observations to be contaminated in the informative variables, default is 0.10.}

\item{scatter_out}{If \code{TRUE}, scattered outliers are generated with the characteristics
specified in \code{s_out_range}, otherwise uniformly distributed outliers are produced with the
specification defined in \code{unif_out_range}.}

\item{p_out_noise}{The number of noise variables in which the contamination is conducted,
see \code{unif_out_range}.}

\item{noise_pct_out}{The proportion of observations to be contaminated in noise variables by replacing
them with uniformly distributed outliers. The contaminated observations differ
from those contaminated in informative variables, default is 0.10.}

\item{unif_out_range}{Optional argument. You can change the interval of an uniform distribution
to randomly generated outliers in \code{[min1,max1] or [min2,max2]}. The specification hase to be in the list,
default is \code{list(min1=-12,max1=-6,min2=6,max2=12)}.}

\item{mu_grp_range}{Optional argument, see references. Default is \code{list(min1=-6,max1=-3,min2=3, max2=6)}.}

\item{s_out_range}{Optional argument, see references. Default is \code{list(min=3,max=10)}.}

\item{rho_grp_range}{Optional argument, see references. Default is \code{list(min=0.1,max=0.9)}.}
}
\value{
\item{x}{A data matrix of a synthetic dataset.}
\item{y}{An integer vector corresponding to a  group membership before contamination.}
\item{lb}{An integer vector with group labels and outlier labels denoted by 0.}
\item{lbout}{An integer vector with group lables and labels for outlier in informative variables (0)
and noise variables (the number of groups+1).}
}
\description{
This function generates a synthetic dataset for the purpose of data clustering (Brodinova et al. (????))
in case of both outliers and noise variables are present.
}
\details{
Groups are generated in the first \code{p_inf} informative variables with various characteristics
following Gaussian models. The groups have different mean vectors and covariance matrix which is additionally
randomly rotated. If uninformative variables are required, \code{p_noise} noise variables
are generated following uniform distributions and added to an informative part.
Two types of outliers - scattered and uniformly distributed- are considered to contaminated data.
The outiers can be placed either in the informative or uninformative part.
}
\examples{
# Generate 3 groups of equal sizes in the first 50 variables with 10\% of
# scatter outliers in all 50 informative variables, and
# 10\% of uniformly distributed outliers in 75 noise variables.

d <- SimData(size_grp=c(40,40,40),p_inf=50,
p_noise=750,p_out_noise=75)

# group membership with outliers in 0 group
table(d$lb)

# scatter outliers in 0 group and uniformly distributed outliers in 4 group
table(d$lbout)

}
\references{
S. Brodinova, P. Filzmoser, T. Ortner, C. Breiteneder, M. Zaharieva. Robust and sparse k-means clustering for
high-dimensional data, 2017.
}
\author{
Sarka Brodinova <sarka.brodinova@tuwien.ac.at>
}
